Create ImageProcessing_Main.py
Created to assist in running ImageProcessing_Program.py. This will take inputs from the user and allow them to press buttons within the Napari interface to actively add layers that coincide with the filters that are used in processing the image. This will make it more accessible and more fine-tuning of the image processing process. Currently, there is an actively working Adaptive Histogram Equalization button and Rescaled Intensity buttons that work on the base image layer.